+++ NEWS.md
@@ -3,7 +3,8 @@
 ## Durga 1.0.0.9000 (development version)
 
 * Added argument violin.params to DurgaPlot to provide additional control over violin appearance.
-* Fixed bug in DurgaBrackets; drawing many brackets sometimes resulted in overlapping brackets.
+* Bug fix in DurgaPlot: violin.fill = TRUE now correctly does not fill violins, rather than white fill.
+* Bug fix in DurgaBrackets; drawing many brackets sometimes resulted in overlapping brackets.
 
 ## Durga 1.0