+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Durga 1.0.0.9000 (development version)
 
+* Improvements to vignette and online help.
 * Added argument `violin.params` to `DurgaPlot` to provide additional control over violin appearance.
 * Bug fix in `DurgaPlot`: `violin.fill = TRUE` now correctly does not fill violins, rather than white fill.
 * Bug fix in `DurgaBrackets`; drawing many brackets sometimes resulted in overlapping brackets.