+++ NEWS.md
@@ -4,7 +4,7 @@
 
 * Improvements to vignette and online help.
 * Added argument `violin.params` to `DurgaPlot` to provide additional control over violin appearance.
-* Bug fix in `DurgaPlot`: `violin.fill = TRUE` now correctly does not fill violins, rather than white fill.
+* Bug fix in `DurgaPlot`: `violin.fill = FALSE` now correctly does not fill violins, rather than white fill.
 * Bug fix in `DurgaBrackets`; drawing many brackets sometimes resulted in overlapping brackets.
 
 ## Durga 1.0