+++ NEWS.md
@@ -6,6 +6,7 @@
 * Added argument `violin.params` to `DurgaPlot` to provide additional control over violin appearance.
 * Bug fix in `DurgaPlot`: `violin.fill = FALSE` now correctly does not fill violins, rather than white fill.
 * Bug fix in `DurgaBrackets`; drawing many brackets sometimes resulted in overlapping brackets.
+* Extend tick marks on right axis to cover CI when plotting standardised effect size on the right.
 
 ## Durga 1.0