+++ NEWS.md
@@ -8,7 +8,7 @@
 * Bug fix in `DurgaPlot`: `violin.fill = FALSE` now correctly does not fill violins, rather than white fill.
 * Bug fix in `DurgaBrackets`; drawing many brackets sometimes resulted in overlapping brackets.
 * Extend tick marks on right axis to cover CI when plotting standardised effect size on the right.
-* Fixed bug in plotting violins for pathological data; failed with error "Error in xy.coords(x, y, setLab = FALSE) : 'x' and 'y' lengths differ"
+* Fixed bug in plotting violins for pathological data; previously failed with error "Error in xy.coords(x, y, setLab = FALSE) : 'x' and 'y' lengths differ"
 
 ## Durga 1.0